Suggestion/backendAPIs/requirement.txt
flask
pymongo
requests
flask_jwt_extended
flask_sqlalchemy
flask-migrate
uuid
flask-cors
python-dotenv
##